// src/modules/party-admissions/enums/admission-step.enum.ts
export enum AdmissionStepCode {
    INITIAL_SUBMISSION = 'INITIAL_SUBMISSION',
    COMMITTEE_REVIEW = 'COMMITTEE_REVIEW',
    DEPUTY_SECRETARY_CONTENT_APPROVAL = 'DEPUTY_SECRETARY_CONTENT_APPROVAL',
    BACKGROUND_VERIFICATION = 'BACKGROUND_VERIFICATION',
    UNION_FEEDBACK_RESOLUTION = 'UNION_FEEDBACK_RESOLUTION',
    RED_STAMP_FINAL_CHECK = 'RED_STAMP_FINAL_CHECK',
    FINAL_SECRETARY_APPROVAL = 'FINAL_SECRETARY_APPROVAL',
    COMPLETED = 'COMPLETED',
    REJECTED = 'REJECTED',
  }